2.1 and 2.2
Added controller that turns on and pauses video on tv. added video on tv that can only be played when controller is pressed. Right hand has no ray anymore I have to get close to object to pick it up. When on object there is a vibration. and when object is picked up there is clicking sound